+++ src/routes/graphql/schema/queries.ts
@@ -1,7 +1,7 @@
 import { GraphQLObjectType, GraphQLList, GraphQLNonNull } from 'graphql';
 import { member, post, profile, user, memberTypeId } from './graphqlTypes.js';
 import { PrismaClient } from '@prisma/client';
-import { UUIDType } from './types/uuid.js';
+import { UUIDType } from '../types/uuid.js';
 
 export const query = new GraphQLObjectType({
   name: 'Query',
