+++ src/routes/graphql/schema/schema.ts
@@ -1,0 +1,9 @@
+import { GraphQLSchema } from 'graphql';
+import { mutation } from './mutations.js';
+import { query } from './queries.js';
+
+export const schema = new GraphQLSchema({
+  query,
+  mutation,
+  // types: [member, user, post, profile],
+});
